backend/requirements.txt
# SmartCity InsightHub Backend - Python 3.11 Requirements
# Compatible with Python 3.11.x

# ============================================================
# FastAPI Framework & ASGI Server
# ============================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ============================================================
# Database - PostgreSQL with pgvector support
# ============================================================
psycopg2-binary==2.9.9

# ============================================================
# Authentication & Security
# ============================================================
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.1

# ============================================================
# Data Validation & Settings
# ============================================================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ============================================================
# AI/ML - Sentence Transformers for Embeddings (384D vectors)
# ============================================================
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
numpy==1.26.2

# ============================================================
# HTTP Client (for testing/external APIs)
# ============================================================
httpx==0.25.1

# ============================================================
# Testing (Optional)
# ============================================================
pytest==7.4.3
pytest-asyncio==0.21.1
